fix bug - var to json

+++ simplecv/mmdet_v2/myapp/app_nms.py
@@ -62,7 +62,7 @@ def _clean_with_iou(dt, mode="min", thr=0.1):
     for i_set in clustering(nodes, lines):
         vals = [dt[i] for i in i_set]
         best = max(vals, key=lambda x: x["score"])
-        bboxes = np.array([d["xyxy"] for d in vals], dtype=np.float64)
+        bboxes = np.array([d["xyxy"] for d in vals], dtype=np.float32)
         (x1, y1), (x2, y2) = bboxes[:, :2].min(axis=0), bboxes[:, 2:].max(axis=0)
         best["bbox"] = [x1, y1, x2 - x1, y2 - y1]
         best["area"] = (x2 - x1) * (y2 - y1)
@@ -93,7 +93,7 @@ def _clean_with_dist(dt, k=1.5):
     for i_set in clustering(nodes, lines):
         vals = [dt[i] for i in i_set]
         best = max(vals, key=lambda x: x["score"])
-        bboxes = np.array([d["xyxy"] for d in vals], dtype=np.float64)
+        bboxes = np.array([d["xyxy"] for d in vals], dtype=np.float32)
         (x1, y1), (x2, y2) = bboxes[:, :2].min(axis=0), bboxes[:, 2:].max(axis=0)
         best["bbox"] = [x1, y1, x2 - x1, y2 - y1]
         best["area"] = (x2 - x1) * (y2 - y1)
